requirements.txt
absl-py==2.1.0
accelerate==1.4.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
asttokens==3.0.0
async-timeout==4.0.3
attrdict==2.0.1
attrs==25.1.0
auto_tune @ file:///root/selfgz384145884/compiler/lib64/auto_tune-0.1.0-py3-none-any.whl#sha256=5374c539972c460012f38b618f00576e36163f03090c8dc58da9d40608065770
av==14.1.0
binpacking==1.5.2
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
cpm-kernels==1.0.11
crcmod==1.7
cryptography==44.0.1
cycler==0.12.1
dacite==1.9.2
dataclasses-json==0.6.7
dataflow @ file:///root/selfgz384145884/compiler/lib64/dataflow-0.0.1-py3-none-any.whl#sha256=0f5d69abccf46385e2ac04c9bf5a6ebcf8d06f9fce6d55fd756420c029e4970e
datasets==3.2.0
decorator==5.1.1
decord==0.6.0
deepspeed==0.16.4
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dotenv==0.9.9
einops==0.8.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
evalscope==0.17.1
evaluate==0.4.5
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
fire==0.7.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
func_timeout==4.3.5
future==1.0.0
fuzzywuzzy==0.18.0
google-auth==2.40.3
google-genai==1.28.0
gradio==5.4.0
gradio_client==1.4.2
greenlet==3.2.3
grpcio==1.70.0
h11==0.14.0
h5py==3.14.0
hccl @ file:///root/selfgz549975916/hccl/lib64/hccl-0.1.0-py3-none-any.whl#sha256=3a3037a8f2fb0949d6fefd479c26e32bda09fd84f3c2b321291d67726cee754d
hccl_parser @ file:///usr/local/Ascend/ascend-toolkit/8.2.RC1/toolkit/tools/hccl_parser-0.1-py3-none-any.whl#sha256=37541281a74de6ae3f8a15c77e1df8cf899bcbeb391a9a47ecfb582caf695fd8
hf-xet==1.2.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.36.0
human-eval==1.0.3
idna==3.10
imageio==2.37.0
immutabledict==4.2.1
importlib_metadata==8.6.1
ipdb==0.13.13
ipython==8.37.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
jmespath==0.10.0
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1733736026804/work
json5==0.12.0
json_repair==0.48.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langdetect==1.0.9
langsmith==0.4.9
latex2sympy2_extended==1.10.2
Levenshtein==0.27.1
llm_datadist @ file:///root/selfgz384145884/compiler/lib64/llm_datadist-0.0.1-py3-none-any.whl#sha256=87888a41b86749195527b9f468b770540f6e338547c989af902246b5b3c79eff
lxml==6.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mmengine-lite==0.10.7
modelscope==1.28.1
mpi4py @ file:///home/conda/feedstock_root/build_artifacts/mpi4py_1637880456026/work
mpmath==1.3.0
ms-opencompass==0.1.6
ms-vlmeval==0.0.17
ms_swift==3.6.3
msgpack==1.1.0
msobjdump @ file:///usr/local/Ascend/ascend-toolkit/8.2.RC1/toolkit/tools/msobjdump-0.1.0-py3-none-any.whl#sha256=9ecf6bf7680333f7bf138c2fb10a512ed183ceb63b63764e775895b31f72fad3
mteb==1.38.20
multidict==6.1.0
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numpy==1.24.4
omegaconf==2.3.0
op_compile_tool @ file:///root/selfgz384145884/compiler/lib64/op_compile_tool-0.1.0-py3-none-any.whl#sha256=5cb767ab2883d45a9b9e819d763ce35d3dd47a7ce1b9e3738144ba94324313a5
op_gen @ file:///usr/local/Ascend/ascend-toolkit/8.2.RC1/toolkit/tools/op_gen-0.1-py3-none-any.whl#sha256=db2b76ec8aa60b562f51e6e33d125a6ca2338d76d03c19a809c63d587cd5c7f6
op_test_frame @ file:///usr/local/Ascend/ascend-toolkit/8.2.RC1/toolkit/tools/op_test_frame-0.1-py3-none-any.whl#sha256=62fdd9628c60936e8b75cbe387781c24a4591b2675b5564ef8f6ff739cc04450
opc_tool @ file:///root/selfgz384145884/compiler/lib64/opc_tool-0.1.0-py3-none-any.whl#sha256=b5fd9db04670c046f0eba7c97953e72ddac87ebe1c27c10351b0b79d93be7de3
openai==1.98.0
OpenCC==1.1.9
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
oss2==2.19.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.8
plotly==5.24.1
polars==1.32.0
portalocker==3.2.0
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
pydicom==3.0.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
pypinyin==0.55.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
python-multipart==0.0.12
pytrec_eval-terrier==0.5.7
pytz==2025.1
PyYAML==6.0.2
qwen-vl-utils==0.0.10
ragas==0.2.14
rank-bm25==0.2.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rouge==1.0.1
rouge-chinese==1.0.3
rouge_score==0.1.2
rsa==4.9.1
ruff==0.9.7
sacrebleu==2.5.1
safehttpx==0.1.6
safetensors==0.5.2
schedule_search @ file:///root/selfgz384145884/compiler/lib64/schedule_search-0.1.0-py3-none-any.whl#sha256=20dd0cb9d8942ec562335e9fce095f999bf0eb3d4a4b3af9eb716e7c92bb3dc9
scikit-learn @ file:///croot/scikit-learn_1737988768944/work
scipy @ file:///croot/scipy_1737122912893/work/dist/scipy-1.15.1-cp310-cp310-linux_aarch64.whl#sha256=e25165130f80018114593a124fb6b178ceeeacbefe290f6be0d3689edc53f37f
screen==1.0.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
shellingham==1.5.4
show_kernel_debug_data @ file:///usr/local/Ascend/ascend-toolkit/8.2.RC1/toolkit/tools/show_kernel_debug_data-0.1.0-py3-none-any.whl#sha256=33333537bad1e8e80280febb19bbe99d964f092c9e72ab96d0e4e2b41e9559b9
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.42
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.45.3
sty==1.0.6
sympy==1.13.1
tabulate==0.9.0
te @ file:///root/selfgz384145884/compiler/lib64/te-0.4.0-py3-none-any.whl#sha256=7144725159b14419ee8861dd2c510bdb4473117beeab67d8a98cd876202eca47
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tiktoken==0.9.0
timeout-decorator==0.5.0
timm==1.0.15
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.12.0
torch==2.4.0
torch-npu==2.4.0.post2
torchvision==0.19.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.0
transformers-stream-generator==0.0.4
trl==0.15.1
typer==0.15.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
validators==0.35.0
wcwidth==0.2.13
webdataset==1.0.2
websockets==12.0
Werkzeug==3.1.3
word2number==1.1
xlsxwriter==3.2.5
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
